Refactor the dir structure
