+++ public/pages/CreateSnapshotPolicy/components/CronSchedule/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import CronSchedule from "./CronSchedule";
+
+export default CronSchedule;
