+++ public/pages/CreateSnapshotPolicy/components/SnapshotAdvancedSettings/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import SnapshotAdvancedSettings from "./SnapshotAdvancedSettings";
+
+export default SnapshotAdvancedSettings;
