+++ public/pages/CreateSnapshotPolicy/components/SnapshotIndicesRepoInput/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import SnapshotIndicesRepoInput from "./SnapshotIndicesRepoInput";
+
+export default SnapshotIndicesRepoInput;
