+++ public/pages/CreateSnapshotPolicy/containers/CreateSnapshotPolicy/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import CreateSnapshotPolicy from "./CreateSnapshotPolicy";
+
+export default CreateSnapshotPolicy;
