+++ public/pages/CreateSnapshotPolicy/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import CreateSnapshotPolicy from "./containers/CreateSnapshotPolicy";
+
+export default CreateSnapshotPolicy;
