+++ public/pages/Repositories/components/CreateRepositoryFlyout/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import CreateRepositoryFlyout from "./CreateRepositoryFlyout";
+
+export default CreateRepositoryFlyout;
