+++ public/pages/Repositories/containers/Repositories/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import RollupDetails from "./Repositories";
+
+export default RollupDetails;
