+++ public/pages/Repositories/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import Repositories from "./containers/Repositories";
+
+export default Repositories;
