--- public/pages/SnapshotManagement/index.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import SnapshotPolicies from "./containers/SnapshotPolicies";
-import Snapshots from "./containers/Snapshots";
-import SnapshotPolicyDetails from "./containers/SnapshotPolicyDetails";
-import CreateSnapshotPolicy from "./containers/CreateSnapshotPolicy";
-import Repositories from "./containers/Repositories";
-
-export { SnapshotPolicies, Snapshots, SnapshotPolicyDetails, CreateSnapshotPolicy, Repositories };
