+++ public/pages/SnapshotPolicies/containers/SnapshotPolicies/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import SnapshotPolicies from "./SnapshotPolicies";
+
+export default SnapshotPolicies;
