+++ public/pages/SnapshotPolicies/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import SnapshotPolicies from "./containers/SnapshotPolicies";
+
+export default SnapshotPolicies;
