+++ public/pages/SnapshotPolicyDetails/containers/SnapshotPolicyDetails/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import SnapshotPolicyDetails from "./SnapshotPolicyDetails";
+
+export default SnapshotPolicyDetails;
