+++ public/pages/SnapshotPolicyDetails/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import SnapshotPolicyDetails from "./containers/SnapshotPolicyDetails";
+
+export default SnapshotPolicyDetails;
