+++ public/pages/Snapshots/components/CreateSnapshotFlyout/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import CreateSnapshotFlyout from "./CreateSnapshotFlyout";
+
+export default CreateSnapshotFlyout;
