+++ public/pages/Snapshots/components/SnapshotFlyout/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import SnapshotFlyout from "./SnapshotFlyout";
+
+export default SnapshotFlyout;
