+++ public/pages/Snapshots/containers/Snapshots/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import Snapshots from "./Snapshots";
+
+export default Snapshots;
